Update FileSystem interface

+++ include/OTwo/Archives/FileInfo.hpp
@@ -8,7 +8,7 @@ class FileInfo final : public Gx::FileInfo
 public:
     FileInfo() = default;
 
-    FileInfo(const Gx::FileSystemController& parent, const std::string& name, const std::int64_t size, const std::uint16_t index, const std::uint64_t offset) :
+    FileInfo(const Gx::FileSystemController& parent, const std::string& name, const std::size_t size, const std::uint16_t index, const std::uint64_t offset) :
         Gx::FileInfo(parent, name, size),
         m_index(index),
         m_offset(offset)
